+++ src/auth/pages/SettingsPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
-import { Navbar } from "../../shared/components/Navbar";
+import { Navbar } from "../../shared/components/navbar/Navbar";
 import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import { useAuth } from "../context/AuthContext";
