+++ src/features/coal/InstructionLaboratoryPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Navbar } from "../../shared/components/Navbar";
+import { Navbar } from "../../shared/components/navbar/Navbar";
 import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import "./InstructionLaboratoryPage.css";
